Add env variables support and move iframe to separate component
