+++ .gitignore
@@ -2,6 +2,8 @@
 #
 .DS_Store
 
+.env
+
 # Xcode
 #
 build/
