+++ package.json
@@ -30,6 +30,7 @@
     "jest": "^29.2.1",
     "metro-react-native-babel-preset": "0.76.5",
     "prettier": "^2.4.1",
+    "react-native-dotenv": "^3.4.9",
     "react-test-renderer": "18.2.0",
     "typescript": "4.8.4"
   },
