+++ types/env.d.ts
@@ -1,0 +1,3 @@
+declare module '@env' {
+  export const SEGMENTS_PATH: string;
+}
